float -> double

+++ CNN/conv_layer_t.hpp
@@ -6,16 +6,16 @@
 class conv_layer_t: public layer_t
 {
 public:
-	std::vector<tensor_t<float>> filters;  // convolution filter kernels
+	std::vector<tensor_t<double>> filters;  // convolution filter kernels
 	std::vector<tensor_t<gradient_t> > filter_grads;
 	uint16_t stride;
 	uint16_t kernel_size;
-	float pad;
+	double pad;
 	
 	conv_layer_t( uint16_t stride,
 		      uint16_t kernel_size, // Width and height of the kernel.  This much of the lower-right edges of the input will be ignored.
 		      uint16_t kernel_count, // Depth of the output.
-		      float pad,
+		      double pad,
 		      tdsize in_size
 		)
 		:
@@ -29,14 +29,14 @@ public:
 		this->kernel_size = kernel_size;
 		throw_assert(kernel_size >= stride, "Convolution kernel size (" << kernel_size << ") must be >= than stride (" << stride << ").");
 		for ( int a = 0; a < kernel_count; a++ ) {	
-			tensor_t<float> t( kernel_size, kernel_size, in_size.z );
+			tensor_t<double> t( kernel_size, kernel_size, in_size.z );
 
 			int maxval = kernel_size * kernel_size * in_size.z;
 
 			for ( int i = 0; i < kernel_size; i++ )
 				for ( int j = 0; j < kernel_size; j++ )
 					for ( int z = 0; z < in_size.z; z++ )
-						t( i, j, z ) = 1.0f / maxval * rand() / float( RAND_MAX );
+						t( i, j, z ) = 1.0f / maxval * rand() / double( RAND_MAX );
 			filters.push_back( t );
 		}
 		for ( int i = 0; i < kernel_count; i++ )
@@ -98,20 +98,20 @@ public:
 		return map_to_output_impl(x,y, kernel_size, stride, filters.size(), out.size);
 	}
 
-	void activate(const tensor_t<float>& in ) {
+	void activate(const tensor_t<double>& in ) {
 		copy_input(in);
 		for ( uint filter = 0; filter < filters.size(); filter++ ) {
-			tensor_t<float>& filter_data = filters[filter];
+			tensor_t<double>& filter_data = filters[filter];
 			for ( int x = 0; x < out.size.x; x++ ) {
 				for ( int y = 0; y < out.size.y; y++ ) {
 					point_t mapped(x*stride, y*stride, 0);
-					float sum = 0;
+					double sum = 0;
 					for ( int i = 0; i < kernel_size; i++ )
 						for ( int j = 0; j < kernel_size; j++ )
 							for ( int z = 0; z < in.size.z; z++ ) {
-								float f = filter_data( i, j, z );
+								double f = filter_data( i, j, z );
 								
-								float v;
+								double v;
 								if (mapped.x + i >= in.size.x ||
 								    mapped.y + j >= in.size.y) {
 									v = pad;
@@ -131,14 +131,14 @@ public:
 			for ( int i = 0; i < kernel_size; i++ )
 				for ( int j = 0; j < kernel_size; j++ )
 					for ( int z = 0; z < in.size.z; z++ ) {
-						float& w = filters[a].get( i, j, z );
+						double& w = filters[a].get( i, j, z );
 						gradient_t& grad = filter_grads[a].get( i, j, z );
 						w = update_weight( w, grad );
 						update_gradient( grad );
 					}
 	}
 
-	void calc_grads(const tensor_t<float>& grad_next_layer ) {
+	void calc_grads(const tensor_t<double>& grad_next_layer ) {
 		throw_assert(grad_next_layer.size == out.size, "mismatch input size for calc_grads");
 
 		for ( uint k = 0; k < filter_grads.size(); k++ ) 
@@ -151,7 +151,7 @@ public:
 			for ( int y = 0; y < in.size.y; y++ ) {
 				range_t rn = map_to_output( x, y );
 				for ( int z = 0; z < in.size.z; z++ ) {
-					float sum_error = 0;
+					double sum_error = 0;
 					for ( int i = rn.min_x; i <= rn.max_x; i++ ) {
 						int minx = i * stride;
 						for ( int j = rn.min_y; j <= rn.max_y; j++ ) {
@@ -211,7 +211,7 @@ namespace CNNTest{
 		tdsize size(10,10,10);
 		conv_layer_t t1(2, 4, 5, 0, size);
 		conv_layer_t t2(2, 4, 5, 0, size);
-		tensor_t<float> in(size);
+		tensor_t<double> in(size);
 		randomize(in);
 		t1.activate(in);
 		EXPECT_EQ(t1,t1);
